Cleaning up a little
We don't really need a closure function inside of a closure function,
so that isn't really needed.  Also seeing as we're only ever going to
be executing the code for building the page once and it's not going to
be that large, I don't think we need to declare a new function for that.