Final update for tonight, should be functional
Everything looks fine, but I don't have anything to test it with.
There's bound to be bugs and what-not.  Cleaned up the rename page
thing a little to be a little more convention-compliant.  The rename
page itself doesn't need a form element, we're not technically
submitting it anywhere.  Notice around line 98 that we need to get the
new name from the rename page.   Either match that ID or change it to
match whatever ID you use for that textbox.  Same goes for the reason.